cleanup(trace-eap-waterfall): Iterating
