+++ static/app/components/quickTrace/utils.tsx
@@ -1,16 +1,30 @@
 import type {Location, LocationDescriptor} from 'history';
+import moment from 'moment-timezone';
 
+import {getTraceDateTimeRange} from 'sentry/components/events/interfaces/spans/utils';
 import {getEventTimestampInSeconds} from 'sentry/components/events/interfaces/utils';
 import {normalizeDateTimeParams} from 'sentry/components/organizations/pageFilters/parse';
 import {ALL_ACCESS_PROJECTS} from 'sentry/constants/pageFilters';
-import type {Event} from 'sentry/types/event';
+import type {Event, EventTransaction} from 'sentry/types/event';
 import type {Organization} from 'sentry/types/organization';
 import EventView from 'sentry/utils/discover/eventView';
 import {SavedQueryDatasets} from 'sentry/utils/discover/types';
-import {getTraceTimeRangeFromEvent} from 'sentry/utils/performance/quickTrace/utils';
 import {hasDatasetSelector} from 'sentry/views/dashboards/utils';
 import type {TraceViewSources} from 'sentry/views/performance/newTraceDetails/traceHeader/breadcrumbs';
 import {getTraceDetailsUrl} from 'sentry/views/performance/traceDetails/utils';
+
+function isTransaction(event: Event): event is EventTransaction {
+  return event.type === 'transaction';
+}
+
+export function getTraceTimeRangeFromEvent(event: Event): {end: string; start: string} {
+  const start = isTransaction(event)
+    ? event.startTimestamp
+    : moment(event.dateReceived ? event.dateReceived : event.dateCreated).valueOf() /
+      1000;
+  const end = isTransaction(event) ? event.endTimestamp : start;
+  return getTraceDateTimeRange({start, end});
+}
 
 export function generateTraceTarget(
   event: Event,
