+++ static/app/views/performance/newTraceDetails/traceApi/utils.tsx
@@ -1,4 +1,3 @@
-import {isTraceSplitResult} from 'sentry/utils/performance/quickTrace/utils';
 import type {TraceItemDetailsResponse} from 'sentry/views/explore/hooks/useTraceItemDetails';
 import type {OurLogsResponseItem} from 'sentry/views/explore/logs/types';
 import type {TraceRootEventQueryResults} from 'sentry/views/performance/newTraceDetails/traceApi/useTraceRootEvent';
@@ -7,6 +6,7 @@ import {
   isEAPTransaction,
   isRootEvent,
   isTraceNode,
+  isTraceSplitResult,
 } from 'sentry/views/performance/newTraceDetails/traceGuards';
 import type {TraceTree} from 'sentry/views/performance/newTraceDetails/traceModels/traceTree';
